update object pool
Former-commit-id: 20538562f92cf9ab7e68cf815c1d75b77f2abd88

+++ Project/Engine/CCollisionMgr.h
@@ -18,7 +18,7 @@ class CCollisionMgr
 	:public CSingletone<CCollisionMgr>
 {
 private:
-	WORD m_matrix[MAX_LAYER ];
+	UINT m_matrix[MAX_LAYER ];
 	map<ULONGLONG, bool> m_mapColInfo;
 
 public :
